fix: move out extractObjectBetween function that trigger update Capability
